Split relational system views
